+++ src/routes/v2/projects/[projectId]/workflows/[workflowId]/+page.svelte
@@ -34,6 +34,7 @@
 	import TimestampCell from '$lib/components/jobs/TimestampCell.svelte';
 	import { normalizePayload } from 'fractal-components';
 	import TemplateCreateModal from '$lib/components/v2/templates/TemplateCreateModal.svelte';
+	import CompareWorkflowTemplateModal from '$lib/components/v2/workflow/CompareWorkflowTemplateModal.svelte';
 
 	const maxDescriptionLength = 50;
 	const descriptionLengthOffset = 10;
@@ -58,6 +59,8 @@
 	let failedJob = $state();
 	/** @type {JobLogsModal|undefined} */
 	let jobLogsModal = $state();
+	/** @type {CompareWorkflowTemplateModal|undefined} */
+	let compareWorkflowToTemplateModal = $state();
 
 	/** @type {import('$lib/components/common/StandardErrorAlert.svelte').default|undefined} */
 	let workflowErrorAlert = undefined;
@@ -729,6 +732,14 @@
 		await setSelectedWorkflowTask(newWorkflowTask);
 	}
 
+	function openCompareWorkflowTemplateModal() {
+		if (argsSchemaForm?.hasUnsavedChanges()) {
+			toggleArgsUnsavedChangesModal();
+		} else {
+			compareWorkflowToTemplateModal?.show();
+		}
+	}
+
 	onDestroy(() => {
 		clearTimeout(statusWatcherTimer);
 	});
@@ -801,9 +812,9 @@
 	{/if}
 
 	<div class="row">
-		<div class="col-lg-8">
+		<div class="col-lg-7 col-md-12">
 			<div class="row">
-				<div class="col-lg-4 col-md-6">
+				<div class="col-lg-4 col-sm-6">
 					<div class="input-group mb-3">
 						<label for="dataset" class="input-group-text">Dataset</label>
 						<select
@@ -819,7 +830,7 @@
 						</select>
 					</div>
 				</div>
-				<div class="col-lg-8 col-md-12">
+				<div class="col-lg-8 col-md-12 mb-2">
 					{#if selectedSubmittedJob && selectedSubmittedJob.status === 'submitted'}
 						<button class="btn btn-danger" onclick={stopWorkflow}>
 							<i class="bi-stop-circle-fill"></i> Stop workflow
@@ -859,7 +870,7 @@
 			</div>
 		</div>
 
-		<div class="col-lg-4 d-flex justify-content-end align-items-start gap-1">
+		<div class="col-lg-5 col-md-12 mb-3 d-flex justify-content-end align-items-start gap-1">
 			{#if page.data.userInfo.is_superuser}
 				<button
 					class="btn btn-light"
@@ -898,6 +909,16 @@
 			>
 				<i class="bi-info-circle"></i>
 			</button>
+			{#if workflow.template_id}
+				<button
+					class="btn btn-light"
+					onclick={openCompareWorkflowTemplateModal}
+					aria-label="Compare workflow to template"
+					title="Compare workflow to template"
+				>
+					<i class="bi bi-plus-slash-minus"></i>
+				</button>
+			{/if}
 			<button
 				class="btn btn-light"
 				onclick={(e) => {
@@ -1527,6 +1548,8 @@
 <JobLogsModal bind:this={jobLogsModal} />
 
 <TemplateCreateModal bind:this={templateCreateModal} {workflow} />
+
+<CompareWorkflowTemplateModal bind:this={compareWorkflowToTemplateModal} {workflow} />
 
 <style>
 	.run-item {